모델 파일 생성
- 앱 모델 : MartType
- 위젯 모델 : WidgetMartType
수정 후 하나의 파일로 병합
