+++ DontGoMart/Model/MartClosedData.swift
@@ -18,3 +18,5 @@ struct MartCloseData: Identifiable {
     var id = UUID().uuidString
     var title: String
 }
+
+var tasks: [MetaMartsClosedDays] = []
